Add Background Life control APIs and queue worker
